[devtools] storybook: port utils to be reusable
